Fix a few more warnings produced by clang.
Andy Wirtz, in cut_selfcup.cpp there were assignments of type 'int cutsideselfcutposition = -0.5;' which do not make much sense. I changed the variable type to double. Please review the three changes I made.


git-svn-id: svn://jacobi.lnm.mw.tum.de/baci/trunk/ccarat@19559 060e22d8-9d1c-0410-868a-e3a950d63d95

+++ src/drt_cut/cut_selfcut.cpp
@@ -1750,7 +1750,7 @@ void GEO::CUT::SelfCut::WallGmsh(const std::string & name)
       throw std::runtime_error("SelfCut: irregular side");
     }
     file << "){";
-    int cutsideselfcutposition = -0.5;
+    double cutsideselfcutposition = -0.5;
     for (int i = 0; i < cutsidetype; ++i)
     {
       if (i > 0)
@@ -2158,7 +2158,7 @@ void GEO::CUT::SelfCut::SCmgmGmsh(const std::string & name)
       throw std::runtime_error("SelfCut: irregular side");
     }
     file << "){";
-    int cutsideselfcutposition = -0.5;
+    double cutsideselfcutposition = -0.5;
     for (int i = 0; i < cutsidetype; ++i)
     {
       if (i > 0)
@@ -2583,7 +2583,7 @@ void GEO::CUT::SelfCut::ErrorGmsh(const std::string & name, Side & cutside)
       throw std::runtime_error("SelfCut: irregular side");
     }
     file << "){";
-    int cutsideselfcutposition = -0.5;
+    double cutsideselfcutposition = -0.5;
     for (int i = 0; i < cutsidetype; ++i)
     {
       if (i > 0)
